Revert "added database layer and changed project structure"
